Implement full submission process
Adds the ability to create, edit, and submit game submissions. The flow
for this follows a stepper model and users can submit up to 3 categories
per game for a total of 15 max. The list of submissions is also
presented at the end of the flow. Submissions are marked as private
whenever touched, even if no fields have otherwise changed.