Fix deletion of game submissions not working
For some reason, the format in which the user object is sent to the
server uses the BSON type for `_id` instead of string. Despite the ids
looking identical, this was causing an equality assertion to fail when
compared to the `runner` id of a `GameSubmission` (also a BSON type).
Using `user.id` instead of `user._id` fixes this, likely due to some
kind of property alias.

Closes: #21