Update schedule link to appear automatically
The plan moving forward is for games list and schedule to be revealed at
the same time, making the need for two separate links unnecessary. This
change ties the schedule link to the `isGamesListPublic` boolean in
order to prevent a required code change for the future.

Also updates the hotel booking link for 2023.